requirements.txt
asgiref==3.4.1
Django==3.2.16
django-ckeditor==5.6.1
django-js-asset==2.0.0
Pillow==8.4.0
pytz==2022.7
sqlparse==0.4.3
typing-extensions==4.1.1